Extract the handling of functions with var args from visibleCandidates.{h,cpp}
